requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
jedi==0.18.2
kombu==5.2.4
packaging==21.3
parso==0.8.3
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pudb==2022.1.3
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
urwid==2.1.2
urwid-readline==0.13
vine==5.0.0
wcwidth==0.2.5
